# api.yaml
openapi: 3.0.3
info:
  title: ARCHIVE API
  version: 1.0.0
  description: |
    **ARCHIVE-BE** — 개인 생산성 및 회고 관리 백엔드 API.

    Clean Architecture 기반 (Domain → Application → Infrastructure → Presentation).

    ### 인증 방식
    - **Access Token**: `Authorization: Bearer <JWT>` 헤더 (만료 15분)
    - **Refresh Token**: `refresh_token` HttpOnly Cookie (만료 7일, `Secure`, `SameSite=Lax`)

    ### 공통 응답 포맷
    모든 응답은 아래 래퍼를 사용한다.
    ```json
    { "status": "success" | "accepted" | "error", "code": "OK", "data": { ... } }
    ```
    에러 응답은 `data: null`이며 `details[]`가 추가된다.

    ### AI 요약 비동기 처리
    `POST /summaries/generate`는 `202 Accepted`만 반환한다.
    실제 결과는 Celery worker가 처리하며, 클라이언트는 `GET /summaries/{id}/stream` SSE로 완료 신호를 받는다.
servers:
  - url: http://localhost:8000/api/v1
    description: 로컬 개발 서버
  - url: https://api.archive.example.com/api/v1
    description: 운영 서버 (placeholder)
...
security:
  - BearerAuth: []

# ────────────────────────────────────────────────────────────────────────────────
paths:
...
  /github/connection:
    get:
      tags: [github]
      summary: GitHub 연결 상태 + push target 통합 조회
      description: |
        FE가 GitHub 화면 진입 시 한 번에 상태를 파악할 수 있도록 묶어 반환한다.

        - `connected`: GitHub OAuth 연결 + 토큰 유효성 검증 결과
        - `login`: GitHub username (connected=true일 때만)
        - `pushTargetRepositoryId`: 회고 push 대상 저장소 id (없으면 null)
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseConnectionStatus' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
  /github/repositories/available:
    get:
      tags: [github]
      summary: GitHub에서 사용자 저장소 목록 조회
      description: |
        OAuth로 연결된 GitHub 계정의 모든 **public** 저장소 목록을 GitHub API에서 직접 가져온다.
        DB 저장은 하지 않음 (연결 후보 표시 용도).
        - `public_repo` scope 기준
        - 페이지네이션 자동 처리 (최대 2000개)
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseAvailableRepoList' }
        '400':
          $ref: '#/components/responses/BadRequest_400'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '429':
          $ref: '#/components/responses/RateLimited_429'
        '503':
          $ref: '#/components/responses/ServiceUnavailable_503'
      x-error-codes:
        - GITHUB_CONNECTION_NOT_FOUND
        - GITHUB_TOKEN_INVALID
        - GITHUB_RATE_LIMITED
        - GITHUB_API_UNAVAILABLE
  /github/repositories:
    get:
      tags: [github]
      summary: 연결된 저장소 목록 조회
      description: DB에 저장된, 현재 사용자가 연결한 GitHub 저장소 목록.
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseRepositoryList' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'

    post:
      tags: [github]
      summary: 특정 저장소 1개 연결
      description: |
        `githubRepoId`로 식별되는 저장소를 사용자 계정에 연결한다.
        백엔드가 GitHub API에서 해당 저장소 메타데이터를 직접 조회해 저장 (위·변조 방지).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LinkRepositoryRequest' }
            example:
              githubRepoId: 1296269
      responses:
        '201':
          description: 연결 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseRepository' }
        '400':
          $ref: '#/components/responses/BadRequest_400'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '404':
          $ref: '#/components/responses/NotFound_404'
        '409':
          $ref: '#/components/responses/Conflict_409'
        '422': { $ref: '#/components/responses/ValidationError_422' }
        '429':
          $ref: '#/components/responses/RateLimited_429'
      x-error-codes:
        - GITHUB_CONNECTION_NOT_FOUND
        - GITHUB_TOKEN_INVALID
        - GITHUB_REPOSITORY_NOT_FOUND
        - GITHUB_REPOSITORY_ALREADY_LINKED
        - GITHUB_RATE_LIMITED

    delete:
      tags: [github]
      summary: 모든 저장소 연결 해제
      description: 현재 사용자의 모든 GitHub 저장소 연결을 일괄 삭제.
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseEmpty' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
  /github/repositories/sync-all:
    post:
      tags: [github]
      summary: GitHub 저장소 전체 일괄 연결 (1회 sync)
      description: |
        호출 시점의 GitHub 저장소 목록을 모두 가져와 일괄 연결한다 (idempotent upsert).
        이후 새로 생긴 repo는 사용자가 다시 호출해야 추가됨 (자동 갱신 없음).
        이미 연결된 저장소는 유지되며, 메타데이터가 바뀌었다면 갱신된다.
      responses:
        '200':
          description: sync 성공. 응답에는 현재 시점에 연결된 모든 저장소 목록이 담긴다.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseRepositoryList' }
        '400':
          $ref: '#/components/responses/BadRequest_400'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '429':
          $ref: '#/components/responses/RateLimited_429'
        '503':
          $ref: '#/components/responses/ServiceUnavailable_503'
      x-error-codes:
        - GITHUB_CONNECTION_NOT_FOUND
        - GITHUB_TOKEN_INVALID
        - GITHUB_RATE_LIMITED
        - GITHUB_API_UNAVAILABLE
  /github/repositories/{repository_id}:
    parameters:
      - in: path
        name: repository_id
        required: true
        schema: { type: string }
        example: ghrepo_01HXYZ...

    patch:
      tags: [github]
      summary: 저장소 역할 설정 (commit 읽기 on/off)
      description: |
        해당 저장소를 회고 화면의 "오늘의 커밋" 집계 대상으로 포함할지 토글한다.
        `commitReadEnabled=false`이면 `GET /github/commits` 응답에서 제외된다.
        Push 대상 지정은 별개 — `commitReadEnabled=false`인 저장소도 push 대상이 될 수 있다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRepositoryRequest' }
            example:
              commitReadEnabled: false
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseRepository' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '404':
          $ref: '#/components/responses/NotFound_404'
        '422': { $ref: '#/components/responses/ValidationError_422' }
      x-error-codes:
        - GITHUB_REPOSITORY_NOT_FOUND

    delete:
      tags: [github]
      summary: 단일 저장소 연결 해제
      responses:
        '200':
          description: 해제 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseEmpty' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '404':
          $ref: '#/components/responses/NotFound_404'
      x-error-codes:
        - GITHUB_REPOSITORY_NOT_FOUND
  /github/commits:
    get:
      tags: [github]
      summary: 지정 날짜(기본=오늘)의 커밋 모음
      description: |
        사용자의 `commitReadEnabled=true` 저장소들에서 지정 날짜의 커밋을 모아
        시간 내림차순으로 반환한다.

        ### 규칙
        - **타임존**: 항상 `user.timezone` 기준. 별도 `tz` 쿼리 파라미터를 받지 않는다.
        - **author 필터**: 사용자 본인의 GitHub 계정(`login`) 커밋만. 첫 호출 시 `/user`
          API 로 한 번 조회 후 `oauth_connections.provider_login` 에 캐시 (lazy backfill).
          이후 호출은 캐시 사용 — 별도 API 호출 없음.
        - **`date` 생략 시**: `user.timezone` 기준 오늘
        - **범위**: `[date 00:00, date+1 00:00)` 사용자 tz → UTC 변환 후 `since`/`until`
        - **scope 정책**: GitHub OAuth scope 가 `public_repo` 만 요청되므로 **public repo
          만 지원**. private repo 는 `repositories/available` 목록에 나오지 않으며,
          어떻게든 등록되어도 GitHub 가 404 → `failedRepositories` 에 포함.

        ### 에러 분류
        | 상황 | 동작 |
        |---|---|
        | GitHub 미연결 / 빈 토큰 | `GITHUB_CONNECTION_NOT_FOUND` (전체 400) |
        | OAuth 토큰 만료/폐기 | `GITHUB_TOKEN_INVALID` (전체 401) |
        | Rate limit | `GITHUB_RATE_LIMITED` (전체 429) |
        | GitHub 5xx | `GITHUB_API_UNAVAILABLE` (전체 503) |
        | 단일 저장소 404 (삭제/private+scope부족) | 해당 repo 만 skip → `failedRepositories[reason="not_found"]` |
        | 단일 저장소 기타 예외 | 해당 repo 만 skip → `failedRepositories[reason="unknown"]`. 백엔드 logger 에 warning |

        ### 응답
        - `commits`: 시간 내림차순. 비어 있어도 200.
        - `failedRepositories`: 부분 실패한 저장소 목록. FE 는 사용자에게 "X개 repo 에서
          커밋을 가져오지 못함" 식으로 표시 권장.
      parameters:
        - in: query
          name: date
          required: false
          schema: { type: string, format: date, pattern: '^\d{4}-\d{2}-\d{2}$' }
          example: '2026-06-12'
          description: 생략 시 user.timezone 기준 오늘
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseCommitList' }
        '400':
          $ref: '#/components/responses/BadRequest_400'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '422': { $ref: '#/components/responses/ValidationError_422' }
        '429':
          $ref: '#/components/responses/RateLimited_429'
        '503':
          $ref: '#/components/responses/ServiceUnavailable_503'
      x-error-codes:
        - GITHUB_CONNECTION_NOT_FOUND
        - GITHUB_TOKEN_INVALID
        - GITHUB_RATE_LIMITED
        - GITHUB_API_UNAVAILABLE
  /github/retrospectives/push:
    post:
      tags: [github]
      summary: 회고 마크다운을 push target 저장소에 commit/push
      description: |
        사용자가 작성한 회고 마크다운 본문을 push target 저장소의 정해진 경로에 commit한다.

        ### 파일 경로 규칙 (사용자 settings.locale 기반 i18n)
        | periodType | 폴더 (ko/en) | 파일명 (ko/en) |
        |---|---|---|
        | DAILY  | `일간/` / `daily/` | `2026-06-12 회고록.md` / `2026-06-12 retrospective.md` |
        | WEEKLY | `주간/` / `weekly/` | `2026-06 2주차 회고록.md` / `2026-06 week 2 retrospective.md` |
        | MONTHLY | `월간/` / `monthly/` | `2026-06 회고록.md` / `2026-06 retrospective.md` |
        | ANNUAL | `년간/` / `annual/` | `2026 회고록.md` / `2026 retrospective.md` |

        ### periodKey 형식
        - DAILY: `YYYY-MM-DD` (예: `2026-06-12`)
        - WEEKLY: `YYYY-MM-Wn` (n=1~6, majority-day 방식)
        - MONTHLY: `YYYY-MM`
        - ANNUAL: `YYYY`

        ### 동작
        - 같은 경로 파일이 있으면 **덮어쓰기**(update). 없으면 **생성**(create).
        - GitHub commit history가 자체 버전 관리 역할 — 덮어써도 이력 유지
        - 푸시 브랜치는 저장소의 `defaultBranch` 사용
        - 커밋 메시지는 사용자 locale로 자동 생성 (prefix `docs(retro):` 영문 고정)
        - 성공 시 백엔드가 `retrospective_pushes` 테이블에 (user_id, period_type, period_key)
          단위로 push 레코드를 upsert. 이후 `GET /entries`, `GET /entries/{id}`,
          `GET /summaries`, `GET /summaries/{id}` 응답에 `githubPush` 필드로 노출됨.
        - `periodType` 'YEARLY' (entry 측 표기) 가 들어오면 BE 가 'annual' 로 정규화해 저장.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PushRetrospectiveRequest' }
            examples:
              daily_ko:
                summary: 일간 회고 (한국)
                value:
                  periodType: DAILY
                  periodKey: '2026-06-12'
                  contentMarkdown: '# 오늘의 회고\n...'
              weekly:
                summary: 주간 회고
                value:
                  periodType: WEEKLY
                  periodKey: '2026-06-W2'
                  contentMarkdown: '# 이번 주 회고\n...'
              monthly:
                summary: 월간 회고
                value:
                  periodType: MONTHLY
                  periodKey: '2026-06'
                  contentMarkdown: '# 6월 회고\n...'
              annual:
                summary: 년간 회고
                value:
                  periodType: ANNUAL
                  periodKey: '2026'
                  contentMarkdown: '# 2026년 회고\n...'
      responses:
        '201':
          description: push 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponsePushResult' }
        '400':
          $ref: '#/components/responses/BadRequest_400'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '422': { $ref: '#/components/responses/ValidationError_422' }
        '502':
          $ref: '#/components/responses/BadGateway_502'
        '503':
          $ref: '#/components/responses/ServiceUnavailable_503'
      x-error-codes:
        - GITHUB_CONNECTION_NOT_FOUND
        - GITHUB_TOKEN_INVALID
        - GITHUB_PUSH_TARGET_NOT_SET
        - GITHUB_REPOSITORY_NOT_LINKED
        - GITHUB_PUSH_FAILED
        - GITHUB_API_UNAVAILABLE

# ────────────────────────────────────────────────────────────────────────────────
components:
...
  securitySchemes:
  # ════════════════════════════════════════════════════════════════════════════
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '`Authorization: Bearer <access_token>` (만료 15분)'
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly · Secure · SameSite=Lax · Max-Age 7d
    OnboardingCookie:
      type: apiKey
      in: cookie
      name: onboarding_token
      description: |
        OAuth 신규 사용자 온보딩용 임시 토큰. HttpOnly · Secure · SameSite=Lax · Max-Age 30분.
        Redis에 `auth:onboarding:{token} → {provider, provider_user_id, email}` 형태로 적재됨.
        클라이언트가 `POST /auth/oauth/onboarding`에서 사용 후 즉시 삭제된다.

  # ════════════════════════════════════════════════════════════════════════════
...
  responses:
...
    BadRequest_400:
      description: |
        잘못된 요청. 가능한 도메인 코드:
        - `RETRO_SUMMARY_INVALID_STATE` — 요약이 잘못된 상태
        - `AUTH_EMAIL_NOT_VERIFIED` — 이메일 미인증 상태에서 register 시도
        - `AUTH_OAUTH_STATE_INVALID` — OAuth state 불일치
        - `GITHUB_CONNECTION_NOT_FOUND` — GitHub 미연결
        - `GITHUB_PUSH_TARGET_NOT_SET` — push 대상 미설정
        - `GITHUB_REPOSITORY_NOT_LINKED` — push 대상이 연결되지 않은 저장소
        - `AUTH_PASSWORD_RESET_NOT_ALLOWED` — OAuth 전용 계정은 비밀번호 재설정 불가
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: RETRO_SUMMARY_INVALID_STATE
            data: null
            details: []
    Unauthorized_401:
      description: |
        인증 실패. 가능한 코드:
        - `AUTH_TOKEN_EXPIRED` — access token 만료 (**FE 자동 refresh 트리거**)
        - `AUTH_TOKEN_INVALID` — 토큰 누락 / 형식 오류 / 서명 불일치 / type mismatch
        - `AUTH_INVALID_CREDENTIALS` — 이메일/비밀번호 불일치
        - `AUTH_REFRESH_TOKEN_INVALID` / `AUTH_REFRESH_TOKEN_REVOKED` — refresh 실패
        - `AUTH_REFRESH_TOKEN_REUSE_DETECTED` — 폐기된 refresh token 재사용 감지(탈취 의심). 해당 사용자의 모든 세션이 즉시 폐기됨. 사용자에게 보안 알림 권장 후 재로그인 유도.
        - `AUTH_OAUTH_STATE_INVALID` — OAuth state 검증 실패
        - `AUTH_ONBOARDING_TOKEN_INVALID` — onboarding cookie 누락
        - `AUTH_ONBOARDING_TOKEN_EXPIRED` — onboarding TTL(30분) 만료
        - `AUTH_PASSWORD_RESET_TOKEN_INVALID` — 토큰 누락/형식 오류
        - `AUTH_PASSWORD_RESET_TOKEN_EXPIRED` — 토큰 TTL(30분) 만료 또는 이미 사용됨
        - `GITHUB_TOKEN_INVALID` — GitHub OAuth 토큰 만료/폐기 (FE는 GitHub OAuth 재인증 유도)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: AUTH_TOKEN_EXPIRED
            data: null
            details: []
    NotFound_404:
      description: |
        리소스를 찾을 수 없음. 가능한 코드:
        - 도메인별 NOT_FOUND (`TODO_NOT_FOUND`, `JOURNAL_ENTRY_NOT_FOUND`, `RETRO_SUMMARY_NOT_FOUND`, `NOTIFICATION_NOT_FOUND`, `USER_NOT_FOUND`, `GITHUB_REPOSITORY_NOT_FOUND`)
        - `AUTH_SESSION_NOT_FOUND` — 폐기 대상 sessionId 가 본인 소유가 아니거나 이미 만료됨
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: TODO_NOT_FOUND
            data: null
            details: []
    Conflict_409:
      description: |
        상태 충돌 (중복·이미 진행 중 등). 가능한 코드:
        - `USER_EMAIL_DUPLICATED` — 이메일 중복 가입
        - `TODO_ALREADY_COMPLETED` / `TODO_ALREADY_IN_PROGRESS`
        - `JOURNAL_ENTRY_ALREADY_EXISTS`
        - `RETRO_SUMMARY_ALREADY_IN_PROGRESS`
        - `GITHUB_REPOSITORY_ALREADY_LINKED`
        - `AUTH_OAUTH_ACCOUNT_ALREADY_LINKED` — link 시도한 OAuth 계정이 다른 사용자에 이미 연결됨
        - `AUTH_OAUTH_PROVIDER_ALREADY_LINKED` — 현재 사용자가 같은 provider에 다른 계정으로 이미 연결됨
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: USER_EMAIL_DUPLICATED
            data: null
            details: []
    ValidationError_422:
      description: |
        요청 검증 실패. 두 가지 경로:
        - **Pydantic 스키마 에러** → `code: VALIDATION_ERROR`, `details` 에 field별 메시지
        - **도메인 검증 에러** (핸들러가 직접 반환) → 도메인 코드 사용:
          - `AUTH_COUNTRY_INVALID` — ISO 3166-1 미등록 국가, 또는 tz 데이터 없는 국가
          - `AUTH_COUNTRY_TIMEZONE_REQUIRED` — 다중 tz 국가에서 timezone 누락
          - `AUTH_TIMEZONE_INVALID` — IANA 형식 아니거나 해당 국가의 옵션 밖
          - `RETRO_SUMMARY_READINESS_UNSUPPORTED` — readiness 점검은 monthly/annual 만 지원, weekly 는 거부
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: VALIDATION_ERROR
            data: null
            details:
              - field: email
                message: value is not a valid email address
    RateLimited_429:
      description: GitHub API rate limit 초과. 잠시 후 재시도.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: GITHUB_RATE_LIMITED
            data: null
            details: []
    BadGateway_502:
      description: |
        외부 서비스(GitHub) push 실패. sha 충돌 / 권한 / 네트워크 오류 등.
        FE는 사용자에게 "다시 시도" 안내. 자동 retry는 금지(중복 commit 위험).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: GITHUB_PUSH_FAILED
            data: null
            details: []
    ServiceUnavailable_503:
      description: 외부 서비스(GitHub API) 일시적 장애
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: GITHUB_API_UNAVAILABLE
            data: null
            details: []

  # ════════════════════════════════════════════════════════════════════════════
  schemas:
...
    ApiResponseEmpty:
      type: object
      required: [status, code, data]
      properties:
        status: { type: string, enum: [success, accepted] }
        code:   { type: string, example: OK }
        data:   { type: 'null' }
    ErrorResponse:
      type: object
      required: [status, code, data]
      properties:
        status: { type: string, enum: [error] }
        code:
          type: string
          description: 도메인 에러 코드 (예 `AUTH_TOKEN_EXPIRED`, `TODO_NOT_FOUND`, `VALIDATION_ERROR`)
        data: { type: 'null' }
        details:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }

    # ── Enum 정의 ───────────────────────────────────────────────────────────
...
    LinkRepositoryRequest:
      type: object
      required: [githubRepoId]
      properties:
        githubRepoId:
          type: integer
          format: int64
          minimum: 1
          description: GitHub의 numeric repository ID
    RepositoryResponse:
      type: object
      required: [id, githubRepoId, owner, name, fullName, isPrivate, defaultBranch, htmlUrl, commitReadEnabled, createdAt]
      properties:
        id:                { type: string, description: 백엔드에서 발급한 연결 ID (`ghrepo_...`) }
        githubRepoId:      { type: integer, format: int64 }
        owner:             { type: string }
        name:              { type: string }
        fullName:          { type: string, example: octocat/Hello-World }
        isPrivate:         { type: boolean }
        defaultBranch:     { type: string, example: main }
        htmlUrl:           { type: string, format: uri }
        commitReadEnabled:
          type: boolean
          default: true
          description: 회고 화면 "오늘의 커밋" 집계 대상 여부 (PATCH로 토글)
        createdAt:         { type: string, format: date-time }
        updatedAt:         { type: string, format: date-time, nullable: true }
    UpdateRepositoryRequest:
      type: object
      required: [commitReadEnabled]
      properties:
        commitReadEnabled: { type: boolean }
    AvailableRepositoryResponse:
      type: object
      required: [githubRepoId, owner, name, fullName, isPrivate, defaultBranch, htmlUrl]
      properties:
        githubRepoId:   { type: integer, format: int64 }
        owner:          { type: string }
        name:           { type: string }
        fullName:       { type: string }
        isPrivate:      { type: boolean }
        defaultBranch:  { type: string }
        htmlUrl:        { type: string, format: uri }
    ApiResponseRepository:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/RepositoryResponse' }
    ApiResponseRepositoryList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/RepositoryResponse' }
    ApiResponseAvailableRepoList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/AvailableRepositoryResponse' }

    # ── GitHub Connection ──────────────────────────────────────────────────
    ConnectionStatusResponse:
      type: object
      required: [connected]
      properties:
        connected:
          type: boolean
          description: GitHub OAuth 연결 + 토큰 유효성 검증 결과
        login:
          type: string
          nullable: true
          example: octocat
          description: connected=true일 때만 값 존재
        pushTargetRepositoryId:
          type: string
          nullable: true
          example: ghrepo_01HXYZ...
          description: 회고 push 대상 저장소 id. 미설정 시 null.
    ApiResponseConnectionStatus:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ConnectionStatusResponse' }

    # ── GitHub Commits ─────────────────────────────────────────────────────
    CommitResponse:
      type: object
      required: [repositoryId, fullName, sha, message, htmlUrl, author, committedAt]
      properties:
        repositoryId: { type: string, description: 백엔드 저장소 연결 ID }
        fullName:     { type: string, example: octocat/Hello-World }
        sha:          { type: string, example: 8f23c41 }
        message:
          type: string
          description: 커밋 메시지 첫 줄만 (서버에서 200자 자름)
        htmlUrl:      { type: string, format: uri }
        author:       { type: string, description: 작성자 GitHub login }
        committedAt:  { type: string, format: date-time }
    FailedRepositoryResponse:
      type: object
      required: [repositoryId, fullName, reason]
      properties:
        repositoryId: { type: string, description: 백엔드 저장소 연결 ID }
        fullName:     { type: string, example: octocat/Hello-World }
        reason:
          type: string
          enum: [not_found, unknown]
          description: |
            - `not_found`: GitHub 가 404 응답. 저장소 삭제, 또는 private+scope 부족.
            - `unknown`: 예기치 못한 예외. 백엔드 logger 의 `github.commits.unknown_error` 참조.
    CommitListResponse:
      type: object
      required: [commits, failedRepositories]
      properties:
        commits:
          type: array
          items: { $ref: '#/components/schemas/CommitResponse' }
        failedRepositories:
          type: array
          items: { $ref: '#/components/schemas/FailedRepositoryResponse' }
    ApiResponseCommitList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CommitListResponse' }

    # ── GitHub Retrospective Push ──────────────────────────────────────────
    PushRetrospectiveRequest:
      type: object
      required: [periodType, periodKey, contentMarkdown]
      properties:
        periodType:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, ANNUAL]
        periodKey:
          type: string
          description: |
            periodType별 형식 (서버 검증 정규식):
            - DAILY:  `^\d{4}-\d{2}-\d{2}$` (예: `2026-06-12`)
            - WEEKLY: `^\d{4}-\d{2}-W[1-6]$` (예: `2026-06-W2`, majority-day 방식)
            - MONTHLY: `^\d{4}-\d{2}$` (예: `2026-06`)
            - ANNUAL: `^\d{4}$` (예: `2026`)
        contentMarkdown:
          type: string
          minLength: 1
          maxLength: 1000000
          description: 회고 본문 (마크다운). 그대로 파일 내용으로 저장됨.
    PushResultResponse:
      type: object
      required: [commitSha, htmlUrl, path]
      properties:
        commitSha: { type: string, example: 8f23c41abcd... }
        htmlUrl:   { type: string, format: uri, description: GitHub 상 파일 또는 커밋 링크 }
        path:      { type: string, example: 일간/2026-06-12 회고록.md }
    ApiResponsePushResult:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/PushResultResponse' }